# apps/api/src/modules/catalog/room-types/room-types.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpStatus,
  Param,
  Patch,
  Put,
  UseInterceptors,
} from '@nestjs/common';
import { ApiCookieAuth, ApiParam, ApiResponse } from '@nestjs/swagger';
import { ZodResponse } from 'nestjs-zod';
import { ErrorResponseDto } from '../../../common/http/error-response.dto';
import {
  type AuthenticatedAdmin,
  CurrentAdmin,
} from '../../auth/current-admin.decorator';
import { SESSION_COOKIE_NAME } from '../../auth/session-cookie';
import { CatalogChangeInterceptor } from '../common/catalog-change.interceptor';
import { RoomTypeIdParamDto } from './dto/params.schema';
import { PatchRoomTypeDto } from './dto/patch-room-type.schema';
import { ReplaceRoomTypeCategoriesDto } from './dto/replace-room-type-categories.schema';
import type {
  RoomTypeAdmin,
  RoomTypeAdminListResponse,
} from './dto/room-type-admin.schema';
import {
  RoomTypeAdminDto,
  RoomTypeAdminListResponseDto,
} from './dto/room-type-admin.schema';
import { RoomTypesService } from './room-types.service';

@Controller('admin/room-types')
@ApiCookieAuth(SESSION_COOKIE_NAME)
@ApiResponse({
  status: HttpStatus.UNAUTHORIZED,
  description: 'UNAUTHENTICATED',
  type: ErrorResponseDto,
})
@UseInterceptors(CatalogChangeInterceptor)
export class RoomTypesController {
  constructor(private readonly roomTypesService: RoomTypesService) {}

  @Get()
  @ZodResponse({ status: 200, type: RoomTypeAdminListResponseDto })
  async list(): Promise<RoomTypeAdminListResponse> {
    return this.roomTypesService.list();
  }

  @Patch(':id')
  @ApiParam({ name: 'id', type: String, format: 'uuid' })
  @ZodResponse({ status: 200, type: RoomTypeAdminDto })
  @ApiResponse({
    status: HttpStatus.CONFLICT,
    description: 'STALE_REVISION',
    type: ErrorResponseDto,
  })
  async updateName(
    @Param() params: RoomTypeIdParamDto,
    @Body() body: PatchRoomTypeDto,
    @CurrentAdmin() admin: AuthenticatedAdmin,
  ): Promise<RoomTypeAdmin> {
    return this.roomTypesService.updateName(params.id, body, admin.id);
  }

  @Put(':id/categories')
  @ApiParam({ name: 'id', type: String, format: 'uuid' })
  @ZodResponse({ status: 200, type: RoomTypeAdminDto })
  @ApiResponse({
    status: HttpStatus.UNPROCESSABLE_ENTITY,
    description: 'CATEGORY_ARCHIVED | CATEGORY_NOT_FOUND',
    type: ErrorResponseDto,
  })
  @ApiResponse({
    status: HttpStatus.CONFLICT,
    description: 'STALE_REVISION',
    type: ErrorResponseDto,
  })
  async replaceCategories(
    @Param() params: RoomTypeIdParamDto,
    @Body() body: ReplaceRoomTypeCategoriesDto,
    @CurrentAdmin() admin: AuthenticatedAdmin,
  ): Promise<RoomTypeAdmin> {
    return this.roomTypesService.replaceCategories(params.id, body, admin.id);
  }
}